String Arrays!
You can now define arrays of strings, and use the for and foreach loops on them. You can also retrieve items from string arrays, and you can get the exact size of a string array.

To declare a string array, use the following code:

string[] | arrayName | {'item1', 'item2', '\0'}

Where arrayName becomes the name of the string array, and each item between the squiggly braces gets added into the string array.
Also note that a string array MUST end with a zero byte string with nothing else in it ('\0' or "\0"). Initialising a string array with no zero byte string will result in a segmentation fault if any attempt is made to get the length of the string array.

The print and println functions can now also print string arrays, and formatting will later be added to print them with style :)