Better indexing and list creation + null datatype
Added the null datatype, which is usually returned when any __index__ call fails to retrieve the index of a type.
Null acts as a placeholder for when a particular value is missing, or for when something isn't quite filled in yet.

Empty lists can now be initialised with `[]`, and indexing now works quite a bit better.

Soon the __display__ method will be added for use with the print function and other methods that return the raw display of a type.