0.0.2 (stable)
Filled in the __disp__ methods of all four main datatypes, including null.

The amount of items that lists can currently hold seems to have no set limit, but list appending slows down decently when a list gets close to 10000 items because of the way appending currently works.

In the future, list appending will be done with listname.add(newlist), but for now it will be done with the += operator.