0.1.0 (stable) - HUGE update
A new datatype named 'long' has been added, which can be instantiated in Quokka code using any integer ending with `l` or `L`. A long can hold an integer value of up to 20 digits in length, whereas a regular int can only hold an integer value of up to 10 digits in length.
I made a demonstration of the differences between int and long at examples/numbers.q.

Procedures can now be created with the `fun` keyword, and in the future they will be able to take arguments and return values, but for now they simply execute a block of code when called.

A lot of bugs have been fixed in this update, and a lot of datatypes now have new and improved operator methods. :grinning: