0.1.2 - Procedures are slowly becoming functions, and all multi-line statements can be ended with 'end' OR two consecutive blank lines
Functions can now return values with the `ret` keyword, but they can't quite take arguments yet.

When starting a multi-line statement or loop, normally you would end it with the `end` keyword, but this is now optional. You can choose whether to leave two consecutive blank lines or to use the `end` keyword.